% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{closelog}
\alias{closelog}
\title{close log file connection}
\usage{
closelog(verbose = TRUE)
}
\arguments{
\item{verbose}{if print success message}
}
\description{
close log file connection
}
\examples{
\dontrun{
closelog()
}
}
