% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printlog.R
\name{createIfNotExist}
\alias{createIfNotExist}
\title{create diretory or/and file if log file path doesn't exits}
\usage{
createIfNotExist()
}
\description{
create diretory or/and file if log file path doesn't exits
}
