% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailylog.R
\name{dailylog}
\alias{dailylog}
\title{daily log}
\usage{
dailylog(...)
}
\arguments{
\item{...}{the message output to log file}
}
\description{
provide daily log function
}
\details{
expecpt the main log file, other log file will be named with its
last modify day, like `log.2000-01-01`, and with only \code{backou_n} log files in this log directory
}
\examples{
\dontrun{
dailylog('this is a test message')
}
}
