% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getBackupN}
\alias{getBackupN}
\title{get log file backup number}
\usage{
getBackupN()
}
\value{
the number of backup log file which was set by user
}
\description{
get log file backup number
}
\examples{
\dontrun{
getBackupN()
}
}
