% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getMaxSize}
\alias{getMaxSize}
\title{get log file max size}
\usage{
getMaxSize()
}
\value{
the max size of log file which was set by user
}
\description{
get log file max size
}
\examples{
\dontrun{
getMaxBytes()
}
}
