% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printlog.R
\name{printlog}
\alias{printlog}
\title{cat message to log}
\usage{
printlog(..., .level = 0)
}
\arguments{
\item{...}{content to be loged, cause we use `cat` function behind the hood, you can put any
R object here}

\item{.level}{the log message level}
}
\description{
just like `cat`, with a timestamp
}
\examples{
\dontrun{
printlog('this is a test message')
}
}
