% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotatelog.R
\name{rotatelog}
\alias{rotatelog}
\title{ratate log}
\usage{
rotatelog(log_name, msg, max_bytes = 200 * 1000, backup_n = 5)
}
\arguments{
\item{log_name}{the main log file name, all other backup log file will be in
the same directory}

\item{msg}{the message output to log file}

\item{max_bytes}{the max log file size, the unit is bytes}

\item{backup_n}{the max number of log files, include main log file}
}
\description{
ratate log
}
\details{
each log file will be size of \code{max_bytes}
at most, and with only \code{backou_n} log files in this log directory
}
\section{Functions}{
\itemize{
\item \code{rotatelog}: rotate log function
}}

