% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readlog.R
\name{readlog}
\alias{readlog}
\title{read log files}
\usage{
readlog(.time = "today", as_json = TRUE)
}
\arguments{
\item{.time}{the min timestamp in log files}

\item{as_json}{if convert log column to multiple columns}
}
\value{
log data frame
}
\description{
read log files, and convert to data frame
}
\examples{
\dontrun{
readlog('today', TRUE)
}
}
