% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setBackupN}
\alias{setBackupN}
\title{set log file backup number}
\usage{
setBackupN(backup_n, verbose = TRUE)
}
\arguments{
\item{backup_n}{the max number of backup log file}

\item{verbose}{if print success message}
}
\description{
if the backup log file number is more than `backup_n`, the earliest log file
will be removed, and all log file will add 1 like 'log.1' -> 'log.2', the log message
will be output to new log file
}
\examples{
\dontrun{
setBackupN(5)
}
}
