% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setMaxBytes}
\alias{setMaxBytes}
\title{set log file size}
\usage{
setMaxBytes(max_bytes, verbose = TRUE)
}
\arguments{
\item{max_bytes}{the max file size in bytes}

\item{verbose}{if print success message}
}
\description{
if the log file size is larger than `max_bytes`, it will be rename like log.1,
and the log message will be output to new log
}
\examples{
\dontrun{
setMaxBytes(100*1024)
}
}
