% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{setMaxSize}
\alias{setMaxSize}
\title{set log file size}
\usage{
setMaxSize(max_size, units = c("Kb", "b", "Mb", "Gb", "Tb", "Pb"),
  verbose = TRUE)
}
\arguments{
\item{max_size}{the max file size in bytes}

\item{units}{the max size units, should be one of `Kb`, `b`, `Mb`, `Gb`, `Tb`, `Pb`}

\item{verbose}{if print success message}
}
\description{
if the log file size is larger than `max_sizes`, it will be rename like log.1,
and the log message will be output to new log
}
\examples{
\dontrun{
setMaxBytes(100*1024)
}
}
