% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getLogInfo}
\alias{getLogInfo}
\title{get log configuration}
\usage{
getLogInfo()
}
\value{
log name, log level, rotate type, backup number, max size, if convert input to json
}
\description{
get log configuration
}
\examples{
\dontrun{
getLogInfo()
}
}
