% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readlog.R
\name{readlog}
\alias{readlog}
\title{read log files}
\usage{
readlog(as_json = TRUE, .time = "today")
}
\arguments{
\item{as_json}{if convert log column to multiple columns}

\item{.time}{the min timestamp in log files}
}
\value{
log data frame
}
\description{
read log files, and convert to data frame
}
\examples{
\dontrun{
readlog('today', TRUE)
}
}
