% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotatelog.R
\name{rotatelog}
\alias{rotatelog}
\title{ratate log}
\usage{
rotatelog(...)
}
\arguments{
\item{...}{the message output to log file}
}
\description{
provide rotate log function
}
\details{
each log file will be size of \code{max_size}
at most, and with only \code{backou_n} log files in this log directory
}
\examples{
\dontrun{
rotatelog('this is a test message')
}
}
